Division by components
