+++ src/App.jsx
@@ -1,0 +1,16 @@
+import "./App.css";
+import Footer from "./components/Footer";
+import Header from "./components/Header";
+import LandingContent from "./components/LandingContent";
+
+function App() {
+  return (
+    <>
+      <Header />
+      <LandingContent />
+      <Footer />
+    </>
+  );
+}
+
+export default App;
